requirements.txt
Flask==0.12              
gunicorn==19.7.1
line-bot-sdk==1.3.0
Beautifulsoup4 == 4.0